# Do not alloc page offset buffer when dictionary encoded